@@ -3,11 +3,11 @@ apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
 
 android {
-  compileSdkVersion 27
+  compileSdkVersion 29
   defaultConfig {
     applicationId "com.pedro.autopermissions"
     minSdkVersion 16
-    targetSdkVersion 27
+    targetSdkVersion 29
     versionCode 100
     versionName "1.0.0"
   }
@@ -22,6 +22,6 @@ android {
 dependencies {
   implementation fileTree(dir: 'libs', include: ['*.jar'])
   implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-  implementation 'com.android.support:appcompat-v7:27.1.1'
+  implementation 'androidx.appcompat:appcompat:1.1.0'
   implementation project(':library')
 }
