@@ -5,13 +5,13 @@ apply plugin: 'com.github.dcendents.android-maven'
 group = 'com.github.pedroSG94'
 
 android {
-  compileSdkVersion 27
+  compileSdkVersion 29
 
   defaultConfig {
     minSdkVersion 14
-    targetSdkVersion 27
-    versionCode 103
-    versionName "1.0.3"
+    targetSdkVersion 29
+    versionCode 104
+    versionName "1.0.4"
   }
 
   buildTypes {
@@ -25,5 +25,5 @@ android {
 dependencies {
   implementation fileTree(dir: 'libs', include: ['*.jar'])
   implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-  implementation 'com.android.support:appcompat-v7:27.1.1'
+  implementation 'androidx.appcompat:appcompat:1.1.0'
 }
