@@ -1,7 +1,7 @@
 package com.pedro.autopermissions
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.widget.Toast
 import com.pedro.library.AutoPermissions
 import com.pedro.library.AutoPermissionsListener
