@@ -3,7 +3,7 @@ package com.pedro.library
 import android.app.Activity
 import android.content.Context
 import android.content.pm.PackageManager
-import android.support.v4.app.ActivityCompat
+import androidx.core.app.ActivityCompat
 
 
 class AutoPermissions {
